// src/enum.ts
export enum TaskStatus {
  Pending = 'Pending',
  Active = 'Active',
  Completed = 'Completed',
}

export enum TagColor {
  Default = '#ffffff',
  White = '#ffffff',
  Red = '#ff6347',
  Orange = '#ff8c00',
  Yellow = '#fff000',
  Green = '#7cfc00',
  Blue = '#1e90ff',
  Violet = '#8a2be2',
  Grey = '#a9a9a9',
}

export enum InputStatus {
  Add = 'Add',
  Edit = 'Edit',
}